vercel-mongo

+++ src/blogs/routers/handlers/delete-blog.handler.ts
@@ -1,15 +1,21 @@
 import { Response, Request } from "express";
 import { blogsRepository  } from "../../repositories/blogs-repositories";
 import { HttpStatus } from "../../../core/types/http.status";
+import { APIErrorResult } from "../../../core/utils/APIErrorResult";
 
-export const deleteBlogHandler = (req: Request, res: Response) => {
+export async function deleteBlogHandler(req: Request<{id: string}>, res: Response){
+try{
+        const id = req.params.id;
+        const deleteBlog = await blogsRepository.findBlogById(id);
 
-        const id = req.params.id.toString();
-        const delBlog = blogsRepository.deleteBlog(id);
-
-        if(!delBlog){
-            res.sendStatus(HttpStatus.NOT_FOUND);
+        if(!deleteBlog){
+            res.status(HttpStatus.NOT_FOUND).send(
+              APIErrorResult([{field: 'id', message: 'Blog not found' }])
+            );
+            return;
         }
-
+         await blogsRepository.deleteBlog(id);
          res.sendStatus(HttpStatus.NO_CONTENT);
-     }
+} catch (err: any){
+  res.sendStatus(HttpStatus.INTERNAL_SERVER_ERROR);
+}}
